Split monolithic test file; make tests Miri-compatible
Split tests/tests.rs (2967 lines) into focused test files:
- roundtrip.rs: pure zrip compress/decompress roundtrips
- adversarial.rs: corruption and fuzz regression tests
- streaming.rs: FrameEncoder/FrameDecoder tests
- dict.rs: dictionary training and roundtrip tests
- proptest_tests.rs: property-based tests
- interop.rs: C zstd cross-validation (gated #[cfg(not(miri))])

Tests that previously used C zstd just to generate compressed test data
now use zrip::compress instead. True interop tests (compress with one
implementation, decompress with the other) are in interop.rs, entirely
skipped under Miri via the file-level #[cfg(not(miri))].

Proptest modules that call C zstd are individually gated. Pure-zrip
proptests run under Miri.
